requirements.txt
tensorflow
dvc
pandas
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML

-e .